Fix MSVC compiler warnings
reg. unreachable code (C4702)
resp. 'initializing': conversion from 'size_t' to 'int', possible loss of data (C4267)
resp. declaration hides class member (C4458)

+++ src/notation/internal/notationinteraction.cpp
@@ -2637,7 +2637,7 @@ void NotationInteraction::editText(QInputMethodEvent* event)
     String& preeditString = m_editData.preeditString;
 
     // remove preedit string
-    int n = preeditString.size();
+    size_t n = preeditString.size();
     while (n--) {
         if (cursor->movePosition(mu::engraving::TextCursor::MoveOperation::Left)) {
             mu::engraving::TextBlock& curLine = cursor->curLine();
